Handle FF ON/OFF for components
Part 3/3: MAke FF working well
